// old/src/app/models/security/administration/params/store-user-param.ts
export interface StoreUserParam {
    id?: number;
    nombres?: string;
    apellidos?: string;
    email?: string;
    username?: string;
    telefono?: string;
    celular?: string;
    documentoIdentidad?: string;
    roles?: string[];
    noCia?: string;
    keyBodega?: string;
    puntoVentaId?: number;
    usuarioEditor?: string;
}